This code exists to compensate for ancient broken build environments which, as far as I can tell, are no longer relevant.  Modern busybox refuses to build under libc5 (there's a specific test and #error for that), and I'm not sure building against 2.1 kernel headers on Alpha was ever relevant.
I'm happy to put any of this back if anybody can point to a real need for it,
but if so we need to specifically document what environment is being
compensated for.  (And we should quarrantine the build environment code
into one place, anyway.  Maybe "quirks.h" for known compiler and
libc quirks?)